+++ portal/portal.conf
@@ -7,14 +7,14 @@ SERVER_NAME = 'localhost:5000'
 DEBUG = True
 SECRET_KEY = '=.DKwWzDd}!3}6yeAY+WTF#W:zt5msTI7]2`o}Y!ziU!#CYD+;T9JpW$ud|5C_3'
 
-DATABASE = './mrdp/data/app.db'
+DATABASE = './portal/data/app.db'
 
-DATASETS= './mrdp/data/datasets.json'
+DATASETS= './portal/data/datasets.json'
 DATASET_ENDPOINT_ID = '30ed71b0-fc1d-11e5-a700-22000bf2d559'
-DATASET_ENDPOINT_BASE = '/mrdp/catalog/'
+DATASET_ENDPOINT_BASE = '/portal/catalog/'
 
 GRAPH_ENDPOINT_ID = '30ed71b0-fc1d-11e5-a700-22000bf2d559'
-GRAPH_ENDPOINT_BASE = '/mrdp/processed/'
+GRAPH_ENDPOINT_BASE = '/portal/processed/'
 
 GA_CLIENT_ID = '82d3ab26-f8da-42fa-9f19-bcd5e7abb668'
 GA_CLIENT_SECRET = 'QmUvb1lZbzJ6W1UuS29aJHQ0UVYyPD0lOHtle0pSLUVDez5iZkIydUo9N3E8'
@@ -48,8 +48,8 @@ GA_LOGOUT_URI = 'https://auth.globus.org/v2/web/logout'
 # 3. obtain an authorization URL, e.g.
 #
 #        >>> from oauth2client import client
-#        >>> from mrdp import app
-#        >>> from mrdp.utils import basic_auth_header
+#        >>> from portal import app
+#        >>> from portal.utils import basic_auth_header
 #        >>>
 #        >>> config = app.config
 #        >>> flow = client.OAuth2WebServerFlow(  # TODO simplify w/ other OAuth calls?
