+++ setup.cfg
@@ -1,2 +1,2 @@
 [flake8]
-exclude = venv/,venv27/,.venv/,mrdp/__init__.py,resource_server/__init__.py
+exclude = venv/,venv27/,.venv/,portal/__init__.py,resource_server/__init__.py
